+++ examples/paint.rs
@@ -3,7 +3,7 @@ use std::cell::RefCell;
 use std::rc::Rc;
 
 fn main() {
-    let app = app::App::default().with_scheme(app::AppScheme::Gtk);
+    let app = app::App::default().with_scheme(app::Scheme::Gtk);
 
     let mut wind = DoubleWindow::new(100, 100, 800, 600, "RustyPainter");
     let mut frame = Frame::new(5, 5, 790, 590, "");