+++ fltk-sys/src/window.rs
@@ -2582,6 +2582,12 @@ extern "C" {
     pub fn Fl_Window_set_region(self_: *mut Fl_Window, r: *mut ::std::os::raw::c_void);
 }
 extern "C" {
+    pub fn Fl_Window_iconize(self_: *mut Fl_Window);
+}
+extern "C" {
+    pub fn Fl_Window_fullscreen_active(self_: *const Fl_Window) -> ::std::os::raw::c_uint;
+}
+extern "C" {
     pub fn Fl_Window_find_by_handle(handle: *mut ::std::os::raw::c_void) -> *mut Fl_Window;
 }
 extern "C" {
@@ -2971,6 +2977,14 @@ extern "C" {
         r: *mut ::std::os::raw::c_void,
     );
 }
+extern "C" {
+    pub fn Fl_Single_Window_iconize(self_: *mut Fl_Single_Window);
+}
+extern "C" {
+    pub fn Fl_Single_Window_fullscreen_active(
+        self_: *const Fl_Single_Window,
+    ) -> ::std::os::raw::c_uint;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Double_Window {
@@ -3342,6 +3356,14 @@ extern "C" {
         r: *mut ::std::os::raw::c_void,
     );
 }
+extern "C" {
+    pub fn Fl_Double_Window_iconize(self_: *mut Fl_Double_Window);
+}
+extern "C" {
+    pub fn Fl_Double_Window_fullscreen_active(
+        self_: *const Fl_Double_Window,
+    ) -> ::std::os::raw::c_uint;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Menu_Window {
@@ -3677,6 +3699,13 @@ extern "C" {
 extern "C" {
     pub fn Fl_Menu_Window_set_region(self_: *mut Fl_Menu_Window, r: *mut ::std::os::raw::c_void);
 }
+extern "C" {
+    pub fn Fl_Menu_Window_iconize(self_: *mut Fl_Menu_Window);
+}
+extern "C" {
+    pub fn Fl_Menu_Window_fullscreen_active(self_: *const Fl_Menu_Window)
+        -> ::std::os::raw::c_uint;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Gl_Window {
@@ -3993,6 +4022,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Gl_Window_set_region(self_: *mut Fl_Gl_Window, r: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Gl_Window_iconize(self_: *mut Fl_Gl_Window);
+}
+extern "C" {
+    pub fn Fl_Gl_Window_fullscreen_active(self_: *const Fl_Gl_Window) -> ::std::os::raw::c_uint;
 }
 extern "C" {
     pub fn Fl_Gl_Window_flush(self_: *mut Fl_Gl_Window);