+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Changes
 - (None)
 
+## [0.7.26] - 2020-08-01
+### Changes
+- Add alias Scheme to AppScheme.
+- Add WindowExt::fullscreen_active() and WindowExt::iconize().
+- Update FLTK.
+
+
 ## [0.7.25] - 2020-07-30
 ### Changes
 - Fix WindowExt::raw_handle() method on systems where `unsigned long int` is u32.